Refactored connection request methods
Moved the methods to create, decline and remove connections into the Connection model and made the into static methods. I did this because I didn't like the idea of having to do an additional query to find the user and instantiate the model just to update the connection. It seems more efficient to just get the two user IDs and do one query to update the record. This is likely how the user connection data will be coming in via an API request, with just the IDs.