Refactored connections
Put the connections query inside of the Connection model (just seemed like a more logical place for it). Changed the query so that it doesn't use SQL fragments. This is mostly just so you could pass either a user or a user id and still get the same result. The other queries work this way so I thought it would be nice for everything to be consistent.