Refactored remove_connection
Moved the query that finds the connection into a named scope. Refactored the query to not use SQL fragment. Once again, just to make it consistent with the others so we could pass either user records or user ids, nothing to do with aesthetic concerns in this case. The remove_connection method now just uses the new for_users scope and throws an exception either if the record isn't found or can't be destroyed.