Updated GraphQL Codegen config
It was really hard to get this working right! Had to look up a lot of things in the docs in order to figure out how to get stricter types that work with what Nuxt is doing. Also had to add a bunch of config for vue-appolo to import things from the correct places for Nuxt 3. Move a few things around to get it more organized.