Added query to connect with other users
I thought about making this part of the user type but I think eventually it would be a pretty big query that gets several different types of connections. It could even be something that's built by a job and gets put into Redis, so it's probably better to make it its own thing.